sockaddr_native: cast sa_family to int in the from_native switch (cosmo fix)
Cosmopolitan types sa_family_t narrowly, so the wider AF_* case labels trip
GCC -Werror=switch-outside-range. Cast the controlling expression to int (the
family value is compared as an integer regardless). No behavior change.

Fixes the Cosmopolitan (APE) CI job on the Phase B change.

+++ src/sockaddr_native.h
@@ -68,7 +68,9 @@ static inline int kl_sockaddr_from_native(KlSockAddr *out,
                                           socklen_t len) {
     if (!out || !sa) return -1;
 
-    switch (sa->sa_family) {
+    /* Cast to int: some platforms (cosmo) type sa_family narrowly (uint8_t),
+     * which trips GCC -Werror=switch-outside-range on the wider AF_* labels. */
+    switch ((int)sa->sa_family) {
     case AF_INET: {
         const struct sockaddr_in *in = (const struct sockaddr_in *)sa;
         uint8_t ip[4];
